+++ Sources/Controllers/ALKConversationViewController.swift
@@ -43,6 +43,12 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         manager.autocompletionDelegate = self
         return manager
     }()
+    
+    public lazy var autoSuggestionManager: AutoCompleteManager = {
+        let manager = AutoCompleteManager(textView: chatBar.textView, tableview: autoSuggestionView)
+        manager.autocompletionDelegate = self
+        return manager
+    }()
 
     public let autocompletionView: UITableView = {
         let tableview = UITableView(frame: CGRect.zero, style: .plain)
@@ -53,8 +59,24 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         tableview.contentInset = UIEdgeInsets(top: 0, left: -5, bottom: 0, right: 0)
         return tableview
     }()
+    
+    public let autoSuggestionView: UITableView = {
+        let tableview = UITableView(frame: CGRect.zero, style: .plain)
+        tableview.backgroundColor = .white
+        tableview.estimatedRowHeight = 25
+        tableview.rowHeight = UITableView.automaticDimension
+        tableview.separatorStyle = .none
+        tableview.contentInset = UIEdgeInsets(top: 0, left: -5, bottom: 0, right: 0)
+        return tableview
+    }()
 
     open lazy var navigationBar = ALKConversationNavBar(configuration: self.configuration, delegate: self)
+    
+    var suggestionArray: [String] = []
+    var suggestionDict: [[String: Any]] = [[:]]
+    var autoSuggestionApi: String = ""
+    var isAutoSuggestionRichMessage: Bool = false
+    var currentText: String = ""
 
     var contactService: ALContactService!
     let registerUserClientService = ALRegisterUserClientService()
@@ -476,6 +498,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         unsubscribingChannel()
         self.invalidateTimerAndUpdateHeightConstraint()
         self.viewModel.timer.invalidate()
+        isAutoSuggestionRichMessage = false
     }
 
     override func backTapped() {
@@ -588,7 +611,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
     }
 
     private func setupConstraints() {
-        var allViews = [backgroundView, contextTitleView, tableView, autocompletionView, moreBar, chatBar, typingNoticeView, unreadScrollButton, replyMessageView, conversationInfoView]
+        var allViews = [backgroundView, contextTitleView, tableView, autocompletionView, autoSuggestionView, moreBar, chatBar, typingNoticeView, unreadScrollButton, replyMessageView, conversationInfoView]
         if let templateView = templateView {
             allViews.append(templateView)
         }
@@ -639,6 +662,10 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
             .constraint(equalTo: typingNoticeView.topAnchor).isActive = true
         autocompletionView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
         autocompletionView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
+        autoSuggestionView.bottomAnchor
+            .constraint(equalTo: typingNoticeView.topAnchor).isActive = true
+        autoSuggestionView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
+        autoSuggestionView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
 
         typingNoticeViewHeighConstaint = typingNoticeView.heightAnchor.constraint(equalToConstant: 0)
         typingNoticeViewHeighConstaint?.isActive = true
@@ -814,6 +841,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
 
                 weakSelf.chatBar.clear()
                 weakSelf.autocompleteManager.cancelAndHide()
+                weakSelf.autoSuggestionManager.cancelAndHide()
 
                 if let profanityFilter = weakSelf.profanityFilter, profanityFilter.containsRestrictedWords(text: message.string) {
                     let profanityTitle = weakSelf.localizedString(
@@ -1144,9 +1172,19 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         Timer.scheduledTimer(timeInterval: timeInterval, target: self, selector: #selector(invalidateTimerAndUpdateHeightConstraint), userInfo: nil, repeats: false)
     }
     
+    public func syncAutoSuggestionMessage(message: ALMessage?){
+        guard let message = message else { return }
+        if message.isAutoSuggestion(),!ALApplozicSettings.isAgentAppConfigurationEnabled() {
+            setupAutoSuggestion(message)
+        }
+    }
+    
     public func sync(message: ALMessage) {
         /// Return if message is sent by loggedin user
         guard !message.isSentMessage() else { return }
+        if message.isAutoSuggestion(),!ALApplozicSettings.isAgentAppConfigurationEnabled() {
+            setupAutoSuggestion(message)
+        }
         guard !viewModel.isOpenGroup else {
             viewModel.syncOpenGroup(message: message)
             return
@@ -2318,6 +2356,7 @@ extension ALKConversationViewController: ALKConversationViewModelDelegate {
         // Clear reply message and the view
         viewModel.clearSelectedMessageToReply()
         hideReplyMessageView()
+        clearAutoSuggestionData()
     }
 
     public func updateTyingStatus(status: Bool, userId: String) {