+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
-##Unreleased
+## [1.1.8] 2023-09-26
+- Added Support For Auto Suggestions Rich Message
 - Added custom input field rich message support in IOS SDK
 - Fixed Trial Period Alert closable issue.
 - Fixed the delete conversation showing the empty screen.  
 - Added support for XCode 15
+- Fixed kommunicate.io showing in iOS 17 even when it is disabled
 
 ## [1.1.7] 2023-09-07
 - Added support for Sending GIF from device
@@ -21,7 +23,6 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 ## [1.1.5] 2023-08-23
 - Crash fix for agent app while updating it to latest ios SDK
 - Added validation and default value for form type rich message
-
 
 ## [1.1.4] 2023-08-11
 - Improved UI of multiple language selection & make it similar to android